+++ src/components/ui/index.ts
@@ -4,3 +4,4 @@ export * from './BackgroundSlideshow';
 export * from './CornerInfoCard';
 export * from './TrendBadge';
 export * from './KpiStrip';
+export * from './DevCornerSlideHero/DevCornerSlideHero';