+++ src/types/index.ts
@@ -4,4 +4,5 @@ export * from './common/TVRoute';
 export * from './common/WidgetConfig';
 export * from './api/ApiResponse';
 export * from './jira';
+export * from './github';
 export * from './charts';